feat: pass through functionality from react-responsive and remove custom Responsive components
BREAKING CHANGE: Custom Responsive components have
been removed in favor of passing through components
and functionality from react-responsive directly.

Components that have been removed:
- `ExtraSmall`
- `Small`
- `Medium`
- `Large`
- `ExtraLarge`
- `ExtraExtraLarge`
- `LargerThanExtraSmall`

New passthroughs:
- `MediaQuery`
- `useMediaQuery`
- `ResponsiveContext`

Please refer to
[react-responsive](https://www.npmjs.com/package/react-responsive)
documentation for more information.
